Refactor trading leaderboard
