+++ apps/frontend/src/app/5_pages/LeaderboardPage/components/Leaderboard/components/Badges/Badges.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 
 import { Badge } from '@sovryn/ui';
 
-import { User } from '../Leaderboard.types';
+import { User } from '../../Leaderboard.types';
 import { getBadgeDetails } from './Badges.utils';
 
 type BadgesProps = {
