+++ apps/frontend/src/app/5_pages/LeaderboardPage/components/Leaderboard/hooks/useGetConnectWalletMessage.tsx
@@ -1,0 +1,30 @@
+import React, { useMemo } from 'react';
+
+import { t } from 'i18next';
+
+import { Button } from '@sovryn/ui';
+
+import { useWalletConnect } from '../../../../../../hooks';
+import { translations } from '../../../../../../locales/i18n';
+
+export const useGetConnectWalletMessage = () => {
+  const { connectWallet } = useWalletConnect();
+
+  const connectWalletMessage = useMemo(
+    () => (
+      <div className="flex items-center justify-center bg-gray-70 lg:bg-gray-90 px-2.5 py-3 lg:p-0 rounded">
+        <div className="text-xs font-medium italic mr-4">
+          {t(translations.leaderboardPage.tables.connectWalletText)}
+        </div>
+        <Button
+          text={t(translations.leaderboardPage.tables.connectWalletCta)}
+          onClick={connectWallet}
+          className="text-sm font-medium"
+        />
+      </div>
+    ),
+    [connectWallet],
+  );
+
+  return connectWalletMessage;
+};
